+++ Project.toml
@@ -22,7 +22,7 @@ CUDA = "^1, ^2"
 DocStringExtensions = "^0.8"
 FFTW = "^1"
 FourierFlows = "^0.6.10"
-JLD2 = "^0.1, ^0.2, ^0.3"
+JLD2 = "^0.1, ^0.2, ^0.3, 0.4"
 Reexport = "^0.2, ^1"
 SpecialFunctions = "^0.10, ^1"
 StaticArrays = "^0.12, ^1"
