### 2021-02-18 #2996 @navidcy
@pdebuyl I'm not sure if dkarrasch would be considered having COI because he has contributed to the repository (see https://github.com/FourierFlows/GeophysicalFlows.jl/commit/a4d42fe90a920e9ab90d6d705afe61c18e553422).
