### 2021-02-19 8616254
some typos + clarifications
