+++ docs/src/stochastic_forcing.md
@@ -23,25 +23,29 @@ numerically implemented then read on; otherwise you can skip this section of the
 
 ## Stochastic Differential Equations (SDEs)
 
-A differential equation in the form:
+A differential equation:
 
 ```math
 	\frac{\mathrm{d} x}{\mathrm{d} t} = f(x) , \quad x(t_0) = 0,
 ```
 
-can also be written in an integral form:
+can also be equivalently written in an integral form:
 
 ```math
 	x(t) = \int_{t_0}^t f(x(s)) \, \mathrm{d} s.
 ```
 
-In a similar manner, a stochastic differential equation
+In a similar manner, a stochastic differential equation (SDE),
 
 ```math
 	\mathrm{d} x = f(x) \, \mathrm{d} t + g(x) \, \mathrm{d} W_t , \quad x(t_0) = 0 ,
 ```
 
-with ``W_t`` a brownian motion or Wiener process.
+with ``W_t`` a Brownian motion or Wiener process, can be written in an integral form as:
+
+```math
+	x(t) = \int_{t_0}^{t} f(x(s)) \, \mathrm{d} s + \int_{t_0}^{t} g(x(s)) \, \mathrm{d} W_s .
+```
 
 !!! tip "Wiener process"
     A Wiener process is a random variable ``W_t`` that depends continuously on ``t \ge 0`` and satisfies the following properties:
@@ -49,16 +53,14 @@ with ``W_t`` a brownian motion or Wiener process.
     2. Stationarity. The statistical distribution of the increment ``W_{t+s} − W_s`` does not depend on  ``s`` (and so is identical in distribution to ``W_t``).
     3. Gaussianity. ``W_t`` is a Gaussian process with mean ``\langle W_t \rangle = 0`` and covariance ``\langle W_t W_s \rangle = \min(t, s)``.
 
-The stochastic differential equation above can be written in an integral form as:
+!!! tip Notation, e.g., ``x_t``
+    It's common to use notation ``x_t`` to denote explicit ``t``-dependence of variable ``x``. Don't confuse here with the other common usage of subscripts denoting partial differentiation.
 
-```math
-	x(t) = \int_{t_0}^{t} f(x(s)) \, \mathrm{d} s + \int_{t_0}^{t} g(x(s)) \, \mathrm{d} W_s .
-```
-
-Of course now, the last integral above is a stochastic integral and there is not a single 
-straight-forward way of computing it -- there are a lot of different ways we can 
-approximate it as a Riemannian sum and each of them leads to a different answer. 
-The two most popular ways for computing such stochastic integrals are:
+The last integral in the integral representation of a SDE expression above is a stochastic integral 
+(it involves a stochastic differential, `` \mathrm{d} W_t``). There is not a single straight-forward 
+way for computing the value of a stochastic integral. The various ways we can approximate the 
+value of a stochastic integral as a Riemannian sum each lead to a different answer. The two most 
+popular ways for computing such stochastic integrals are:
 
 ```math
 \begin{aligned}
@@ -69,18 +71,17 @@ The two most popular ways for computing such stochastic integrals are:
 
 The difference in the two calculi above lies in the point at which we choose to evaluate ``g(x)``:
 we take the start of the time-interval for ${\color{Green} \text{Itô, } t_j}$, while we use
-the mid-point for ${\color{Magenta}\text{Stratonovich, } (t_j+t_{j+1})/2}$. In the case the 
+the mid-point for ${\color{Magenta}\text{Stratonovich, } \tfrac{1}{2}(t_j+t_{j+1})}$. In the case the 
 stochastic noise is additive, i.e., its prefactor ``g`` does not depend on the state ``x_t``,
[…]
