+++ docs/src/stochastic_forcing.md
@@ -54,7 +54,7 @@ with ``W_t`` a Brownian motion or Wiener process, can be written in an integral 
     3. Gaussianity. ``W_t`` is a Gaussian process with mean ``\langle W_t \rangle = 0`` and covariance ``\langle W_t W_s \rangle = \min(t, s)``.
 
 !!! tip Notation, e.g., ``x_t``
-    It's common to use notation ``x_t`` to denote explicit ``t``-dependence of variable ``x``. Don't confuse here with the other common usage of subscripts denoting partial differentiation.
+    It's common to use notation ``x_t`` to denote explicit ``t``-dependence of variable ``x``. Not to be confused with the other common usage of subscripts for denoting partial differentiation.
 
 The last integral in the integral representation of a SDE expression above is a stochastic integral 
 (it involves a stochastic differential, `` \mathrm{d} W_t``). There is not a single straight-forward 
