### 2021-02-22 #3053 @whedon
[REVIEW]: GeophysicalFlows.jl: Solvers for geophysical fluid dynamics problems in periodic domains on CPUs & GPUs
**Submitting author:** @navidcy (<a href="http://orcid.org/0000-0002-8149-4094">Navid C. Constantinou</a>)
**Repository:** <a href="https://github.com/FourierFlows/GeophysicalFlows.jl" target ="_blank">https://github.com/FourierFlows/GeophysicalFlows.jl</a>
**Version:** v0.12.1
