### 2021-02-25 08635d2
minor tweaks in figures
