### 2021-02-25 4978b1d
remove dupicate example_path
