+++ docs/make.jl
@@ -32,7 +32,6 @@ examples = [
 ]
 
 for example in examples
-  example_filepath = joinpath(EXAMPLES_DIR, example)
   withenv("GITHUB_REPOSITORY" => "FourierFlows/GeophysicalFlowsDocumentation") do
     example_filepath = joinpath(EXAMPLES_DIR, example)
     Literate.markdown(example_filepath, OUTPUT_DIR, documenter=true)
