### 2021-02-25 f2bf5eb
generalize initialization of ψh
