### 2021-02-26 #3053 @eviatarbach
Before getting into the rest of the review, I ran into some installation issues on older versions of Julia. On Julia 1.5.3 the installation works fine.

On Julia 1.4.1 (less than a year old), the error https://github.com/FourierFlows/GeophysicalFlows.jl/issues/128 occurs upon first import. I understand this was fixed in GeophysicalFlows.jl version 0.8.1, but that version requires Julia 1.5.0, meaning that the package installation is broken on Julia 1.4.1 (and I assume all versions of Julia between 1.4.0 and 1.5.0). I suggest that either a fix is backported or the package be marked as incompatible with these versions.

I also tried installing it on Julia 1.3.1, which results in an error about unsatisfiable requirements. This one is perhaps not as important because it gives an explicit error at installation time. However, I think it would be good to determine which versions of Julia this package is compatible with at present (especially versions since Julia 1.0.5, the LTS), and state this in the repository README.
