### 2021-02-28 fefb192
drop compat requirement for julia 1.5->1.4 in ^0.8
