+++ README.md
@@ -1,11 +1,8 @@
 # GeophysicalFlows.jl
 
 <p align="left">
-    <a href="https://travis-ci.com/FourierFlows/GeophysicalFlows.jl">
-        <img alt="Build Status for CPU" src="https://img.shields.io/travis/com/FourierFlows/GeophysicalFlows.jl/master?label=CPU&logo=travis&logoColor=white&style=flat-square">
-    </a>
-    <a href="https://gitlab.com/JuliaGPU/GeophysicalFlows-jl/commits/master">
-      <img alt="Build Status for GPU" src="https://img.shields.io/gitlab/pipeline/JuliaGPU/GeophysicalFlows-jl/master?label=GPU&logo=gitlab&logoColor=white&style=flat-square">
+    <a href="https://buildkite.com/julialang/geophysicalflows-dot-jl">
+        <img alt="Buildkite CPU+GPU build status" src="https://img.shields.io/buildkite/4d921fc17b95341ea5477fb62df0e6d9364b61b154e050a123/master?logo=buildkite&label=Buildkite%20CPU%2BGPU">
     </a>
     <a href="https://ci.appveyor.com/project/navidcy/geophysicalflows-jl">
         <img alt="Build Status for Window" src="https://img.shields.io/appveyor/ci/navidcy/geophysicalflows-jl/master?label=Window&logo=appveyor&logoColor=white&style=flat-square">
