+++ .buildkite/pipeline.yml
@@ -1,15 +1,15 @@
 env:
-  JULIA_VERSION: "1.5.3"
-  JULIA_MINOR_VERSION: "1.5"
+  JULIA_VERSION: "1.4.2"
+  JULIA_MINOR_VERSION: "1.4"
   GKSwstype: "100" # See: https://github.com/jheinen/GR.jl/issues/278
   SECRET_CODECOV_TOKEN: "Ak2mVTxXnhkPNc096ImDdp7bOc4zGNTqFEEDaGMwAgYPr28g5dyMbslh8B/ad4NQHXVL1MXQ3zrUfGgMBRq+mmqRaAe13FI4Go9uCas6bzdZXE23ExiLzBmqVRNRf8GqEcpGL7BBreohC0cnfI0SVMiIJDCJXX9YsXJtlcpk1glQFMEFI5V6cpFe9K2l5xoUNQ4179ZYoJUMAy/aylQx/UdQuw527FjHQUsi5/dFtWzMqeys0secNa9alLvJCQdIX9OqPjmBYvuIIVXCR7vlZoH8PgXwEj7wbdp8/V31+wlLQI9WePcsJxoOybtLTOlwwfw4jWLAttDZYqnqiLVp3Q==;U2FsdGVkX18sNLCManU1B/jI5kh4LhSi69MFXljHSp9yWrN7u5d196K/XrELwb8ksbamyKeHjIvDIopwD55dbw=="
 
 
 steps:
-  - label: "Julia 1.5"
+  - label: "Julia 1.4"
     plugins:
       - JuliaCI/julia#v1:
-          version: 1.5
+          version: 1.4
       - JuliaCI/julia-test#v1: ~
       - JuliaCI/julia-coverage#v1:
           codecov: true
