### 2021-03-02 5b5dcf5
don't test multilaerqg on gpu
