+++ README.md
@@ -43,7 +43,7 @@ All modules provide solvers on two-dimensional domains. We currently provide
 * `TwoDNavierStokes`: the two-dimensional vorticity equation.
 * `BarotropicQG`: the barotropic quasi-geostrophic equation, which generalizes `TwoDNavierStokes` to cases with topography and Coriolis parameters of the form `f = f₀ + βy`.
 * `BarotropicQGQL`: the quasi-linear barotropic quasi-geostrophic equation.
-* `MultilayerQG`: a multi-layer quasi-geostrophic model over topography and with the ability to impose a zonal flow `U_n(y)` in each layer.
+* `MultilayerQG`: a multi-layer quasi-geostrophic model over topography and with the ability to impose a zonal flow `U_n(y)` in each layer. [*Note*: `MultilayerQG` is not tested on GPU; if you want to use `MultilayerQG` on the GPU use GeophysicalFlows.jl v0.11 or later.]
 * `SurfaceQG`: a surface quasi-geostrophic model.
 
 
