### 2021-03-02 #3053 @navidcy
hi @eviatarbach, regarding the issue with Julia v1.4.2, I've pushed a bugfix release v0.8.5 which should be fine.

```
julia> using GeophysicalFlows
[ Info: Precompiling GeophysicalFlows [44ee3b1c-bc02-53fa-8355-8e347616e15e]
[ Info: FourierFlows will use 12 threads
[ Info: FourierFlows will use 12 threads
```

and

```
(@v1.4) pkg> test GeophysicalFlows
    Testing GeophysicalFlows
Status `/private/var/folders/1z/s8gw4q055514jw75ztnclx4h0000gn/T/jl_Mqed4V/Manifest.toml`
  [621f4979] AbstractFFTs v1.0.1
  [79e6a3ab] Adapt v2.4.0
  [56f22d72] Artifacts v1.3.0
  [13072b0f] AxisAlgorithms v1.0.0
  [b99e7846] BinaryProvider v0.5.10
  [fa961155] CEnum v0.4.1
  [052768ef] CUDA v1.3.3
  [d360d2e6] ChainRulesCore v0.9.29
  [944b1d66] CodecZlib v0.7.0
  [34da2185] Compat v3.25.0
  [e66e0078] CompilerSupportLibraries_jll v0.3.4+0
  [a2441757] Coverage v1.2.0
  [c36e975a] CoverageTools v1.2.2
  [864edb3b] DataStructures v0.18.9
  [ffbed154] DocStringExtensions v0.8.3
  [e2ba6199] ExprTools v0.1.3
  [7a1cc6ca] FFTW v1.3.2
  [f5851436] FFTW_jll v3.3.9+7
  [2aec4490] FourierFlows v0.6.8
  [0c68f7d7] GPUArrays v5.2.1
  [61eb1bfa] GPUCompiler v0.6.1
  [44ee3b1c] GeophysicalFlows v0.8.5
  [cd3eb016] HTTP v0.9.5
  [83e8ac13] IniFile v0.5.0
  [1d5cc7b8] IntelOpenMP_jll v2018.0.3+2
  [a98d9a8b] Interpolations v0.13.1
  [033835bb] JLD2 v0.3.3
  [692b3bcd] JLLWrappers v1.2.0
  [682c06a0] JSON v0.21.1
  [929cbde3] LLVM v2.0.0
  [4af54fe1] LazyArtifacts v1.3.0
  [856f044c] MKL_jll v2021.1.1+1
  [1914dd2f] MacroTools v0.5.6
  [739be429] MbedTLS v1.0.3
  [c8ffd9c3] MbedTLS_jll v2.16.8+1
  [872c559c] NNlib v0.7.14
  [ca575930] NetworkOptions v1.2.0
  [6fe1bfb0] OffsetArrays v1.6.1
  [efe28fd5] OpenSpecFun_jll v0.5.3+4
  [bac558e1] OrderedCollections v1.4.0
  [69de0a69] Parsers v1.0.15
  [c84ed2f1] Ratios v0.4.0
  [189a3867] Reexport v0.2.0
  [ae029012] Requires v1.1.2
  [276daf66] SpecialFunctions v1.3.0
  [90137ffa] StaticArrays v0.12.5
  [a759f4b9] TimerOutputs v0.5.8
  [3bb67fe8] TranscodingStreams v0.9.5
  [5c2747f8] URIs v1.2.0
  [efce3f68] WoodburyMatrices v0.5.3
  [83775a58] Zlib_jll v1.2.11+18
  [2a0f44e3] Base64
  [ade2ca70] Dates
  [8bb1440f] DelimitedFiles
  [8ba89e20] Distributed
  [b77e0a4c] InteractiveUtils
  [76f85450] LibGit2
  [8f399da3] Libdl
  [37e2e46d] LinearAlgebra
  [56ddb016] Logging
  [d6f4376e] Markdown
  [a63ad114] Mmap
  [44cfe95a] Pkg
  [de0858da] Printf
  [3fa0cd96] REPL
  [9a3f8284] Random
  [ea8e919c] SHA
  [9e88b42a] Serialization
  [1a1011a3] SharedArrays
  [6462fe0b] Sockets
  [2f01184e] SparseArrays
  [10745b16] Statistics
  [8dfed614] Test
  [cf7118a7] UUIDs
  [4ec0a83e] Unicode
[ Info: FourierFlows will use 12 threads
testing on CPU
Test Summary: | Pass  Total
Utils         |    2      2
Test Summary:    | Pass  Total
TwoDNavierStokes |    9      9
Test Summary: | Pass  Total
BarotropicQG  |   18     18
Test Summary:  | Pass  Total
BarotropicQGQL |   14     14
Test Summary: | Pass  Total
SurfaceQG     |    9      9
testing on GPU
Test Summary: | Pass  Total
Utils         |    2      2
┌ Warning: calls to Base intrinsics might be GPU incompatible
│   exception =
│    You called cos(x::T) where T<:Union{Float32, Float64} in Base.Math at special/trig.jl:100, maybe you intended to call cos(x::Float64) in CUDA at /g/data/v45/nc3020/.julia/packages/CUDA/dZvbp/src/device/intrinsics/math.jl:5 instead?
│    Stacktrace:
│     [1] cos at special/trig.jl:100
│     [2] broadcast_kernel at /g/data/v45/nc3020/.julia/packages/GPUArrays/uaFZh/src/host/broadcast.jl:60
└ @ GPUCompiler /g/data/v45/nc3020/.julia/packages/GPUCompiler/GKp4B/src/irgen.jl:68
Test Summary:    | Pass  Total
TwoDNavierStokes |    9      9
Test Summary: | Pass  Total
BarotropicQG  |   18     18
Test Summary:  | Pass  Total
BarotropicQGQL |   14     14
Test Summary: | Pass  Total
SurfaceQG     |    9      9
Test Summary: | Pass  Total
MultilayerQG  |   14     14
Total test time: 1111.3896864
    Testing GeophysicalFlows tests passed
