### 2021-03-02 #3053 @navidcy
@eviatarbach, regarding supporting Julia's longterm release v1.0.5, we can't be using CUDA.jl (since even [v0.1.0](https://github.com/JuliaGPU/CUDA.jl/blob/1ed8b5144b9c9fbee744bda9ab5745e0da3b93a4/Project.toml#L44) requires Julia v1.3). Unfortunately, the latest GeophysicalFlows.jl version that is supported on Julia v1.0.5 is v0.5.1.
