### 2021-03-02 1fe5dfc
adds info regarding version compatibilities
