### 2021-03-02 #3053 @navidcy
I have added remark regarding version compatibilities with Julia's releases in the readme; see [`1fe5dfc`](https://github.com/FourierFlows/GeophysicalFlows.jl/commit/1fe5dfced584ad3494234dd8ca1f09dbd4e61bb3).
