### 2021-03-03 #3053 @pdebuyl
@navidcy typically a JOSS reviews considers a single version of the software. The version listed upon submission is v0.11.3, may I ask what is the relevance of updating earlier versions?
