### 2021-03-03 #3053 @eviatarbach
@navidcy Thank you for addressing this issue! It works now on Julia 1.4.

@pdebuyl I brought up this issue. I thought that this would be important to address since the installation was failing on even quite recent versions of Julia. I apologize if this was outside the scope of the review.
