### 2021-03-03 #3053 @navidcy
@pdebuyl, as @eviatarbach explained I was trying to address their remark above. I felt that adding a note in the readme was indeed very appropriate.
