### 2021-03-03 #3053 @pdebuyl
@eviatarbach this was not out of scope, no problem. I did not get why there was a discussion of older versions of the code but it all makes sense. + There is now a bugfix release and extra information in the README, which is good.
