### 2021-03-04 FourierFlows/GeophysicalFlows.jl#203 @ranocha
Launching the example "2D dcaying turbulence" in binder does not work
I get
![image](https://user-images.githubusercontent.com/12693098/109988712-17f3dc00-7d08-11eb-90c0-4022bc91e305.png)
