### 2021-03-04 #3053 @ranocha
Thanks for creating this nice package (and the basic functionality in FourierFlows.jl), @navidcy et al. Please find below some comments and suggestions.

- [ ] You should consider reformatting the license file `LICENSE.md` such that GitHub recognizes the license type automatically, cf. https://opensource.org/licenses/MIT
- [ ] It would be good to have a top-level CONTRIBUTING file (mentioned also in the docs) so users see a link to it when they open a PR. It would also be good to note if you prefer users create a GitHub issue or use some other means for reporting problems. Related issue: https://github.com/FourierFlows/GeophysicalFlows.jl/issues/180
- [ ] The currently latest release is v0.8.5 but the README tells people to cite v0.11.4 - that's really confusing and weird.
- [ ] There is no explicit section "Statement of need" required for review. It would be good to structure the paper a bit more, e.g. by including sections like "Features" and "Related research and software" (although some related information is contained in the manuscript already).
- [ ] Would it be possible to leverage ApproxFun.jl (and/or related Matlab tools) for these problems?
- [ ] What is the reason for implementing the time stepping algorithms on your own instead of using something like OrdinaryDiffEq.jl?
- [ ] How do you think about relations to Oceananigans.jl?
- [ ] From reading the manuscript and the README, it's not clear to me what kind of scalability to expect. Is it correct that this package is restricted to using a single CPU or GPU, mostly in two spatial dimensions? This should be made more clear.
- [ ] What about integration of functionality from PencilFFTs.jl for parallelism? Would something like this be possible?
- [ ] I could not launch some examples in Binder as described in the documentation, see https://github.com/FourierFlows/GeophysicalFlows.jl/issues/203
- [ ] I also tried to run the first few examples from the documentation on the GPU but got some errors like `ERROR: scalar getindex is disallowed`. It would be really nice to point the reader explicitly to some GPU examples or (even better) to make the CPU examples compatible with running them on the GPU.
