### 2021-03-04 #3053 @pdebuyl
Thank you @ranocha for the review.

The statement of need is supposed to be checked on submission, probably it wasn't because of the non "master" branch for the paper.
