### 2021-03-04 7f86d98
adds Contributor's guide
