+++ docs/src/contributing.md
@@ -1,0 +1,122 @@
+# Contributors Guide
+
+This is a short guide for potential GeophysicalFlows.jl contributors.
+
+Please feel free to ask us questions and chat, either by raising an [issue](https://github.com/FourierFlows/GeophysicalFlows.jl/issues) or starting a [discussion](https://github.com/FourierFlows/GeophysicalFlows.jl/discussions).
+
+We follow the [ColPrac guide](https://github.com/SciML/ColPrac) for collaborative practices. 
+New contributors should make sure to read that guide.
+
+## What can I do?
+
+* Tackle an existing [issue](https://github.com/FourierFlows/GeophysicalFlows.jl/issues).
+
+* Try to run your favorite GeophysicalFlows.jl module and play around with it to simulate 
+  your particular favorite setup. If you run into any problems or find it difficult
+  to use, modify, or understand, please [open an issue](https://github.com/FourierFlows/GeophysicalFlows.jl/issues)!
+
+* Write up an example or tutorial on how to do something useful with Oceananigans, like how 
+  to set up a new physical configuration.
+
+* Improve documentation, docstrings, or comments if you found something is hard to use.
+
+* Implement a new feature (e.g., a new diagnostic into a module).
+
+* Implement a new module from scratch to solve your favorite partial differential equation with
+  periodic boundary conditions.
+
+If you're interested in working on something, let us know by commenting on
+existing issues or by opening a new issue if. This is to make sure no one else
+is working on the same issue and so we can help and guide you in case there
+is anything you need to know beforehand.
+
+## Ground Rules
+
+* Each pull request should consist of a logical collection of changes. You can
+  include multiple bug fixes in a single pull request, but they should be related.
+  For unrelated changes, please submit multiple pull requests.
+* Do not commit changes to files that are irrelevant to your feature or bugfix
+  (e.g., `.gitignore`).
+* Be willing to accept criticism and work on improving your code; we don't want
+  to break other users' code, so care must be taken not to introduce bugs. We
+  discuss pull requests and keep working on them until we believe we've done a
+  good job.
+* Be aware that the pull request review process is not immediate, and is
+  generally proportional to the size of the pull request.
+
+## Reporting a bug
+
+The easiest way to get involved is to report issues you encounter when using
+Oceananigans or by requesting something you think is missing.
+
+* Head over to the [issues](https://github.com/CLiMA/Oceananigans.jl/issues) page.
+* Search to see if your issue already exists or has even been solved previously.
+* If you indeed have a new issue or request, click the "New Issue" button.
+* Please be as specific as possible. Include the version of the code you were using, as
+  well as what operating system you are running. The output of Julia's `versioninfo()`
+  and `] status` is helpful to include. If possible, include complete, minimal example
+  code that reproduces the problem.
+
+## Setting up your development environment
+
+* Install [Julia](https://julialang.org/) on your system.
+* Install git on your system if it is not already there (install XCode command line tools on
+  a Mac or git bash on Windows).
+* Login to your GitHub account and make a fork of the
+  [GeophysicalFlows.jl repository](https://github.com/FourierFlows/GeophysicalFlows.jl) by
+  clicking the "Fork" button.
+* Clone your fork of the Oceananigans repository (in terminal on Mac/Linux or git shell/
+  GUI on Windows) in the location you'd like to keep it.
+  ```
+  git clone https://github.com/your-user-name/GeophysicalFlows.jl.git
+  ```
+* Navigate to that folder in the terminal or in Anaconda Prompt if you're on Windows.
+* Connect your repository to the upstream (main project).
+  ```
+  git remote add oceananigans https://github.com/FourierFlows/GeophysicalFlows.jl.git
+  ```
[…]
