+++ CONTRIBUTING.md
@@ -1,0 +1,40 @@
+# Contributors Guide
+
+Thank you for considering contributing to GeophysicalFlows.jl! 
+
+Please feel free to ask us questions and chat with us at any time if you're
+unsure about anything.
+
+[issue](https://github.com/FourierFlows/GeophysicalFlows.jl/issues) or starting a [discussion](https://github.com/FourierFlows/GeophysicalFlows.jl/discussions).
+
+Best way to get in touch is to either just open a GitHub [issue](https://github.com/FourierFlows/GeophysicalFlows.jl/issues) 
+(don't be shy!) or starting a [discussion](https://github.com/FourierFlows/GeophysicalFlows.jl/discussions).
+
+We follow the [ColPrac guide](https://github.com/SciML/ColPrac) for collaborative
+practices. New contributors should make sure to read that guide.
+
+## What can I do?
+
+* Tackle an existing issue. We keep a list of [good first issues](https://github.com/climate-machine/Oceananigans.jl/issues?q=is%3Aissue+is%3Aopen+label%3A%22good+first+issue%22)
+  that are self-contained and suitable for a newcomer to try and work on.
+
+* Try to run Oceananigans and play around with it to simulate your favorite
+  fluids and ocean physics. If you run into any bugs/problems or find it difficult
+  to use or understand, please open an issue!
+
+* Write up an example or tutorial on how to do something useful with
+  Oceananigans, like how to set up a new physical configuration.
+
+* Improve documentation or comments if you found something hard to use.
+
+* Implement a new feature if you need it to use Oceananigans.
+
+If you're interested in working on something, let us know by commenting on
+existing issues or by opening a new issue. This is to make sure no one else
+is working on the same issue and so we can help and guide you in case there
+is anything you need to know beforehand.
+
+We also hang out on the #oceananigans channel on Julia Slack, which is a great
+place to discuss anything Oceananigans-related, especially contributions! To
+join the Julia Slack, go to [https://julialang.org/slack/](https://julialang.org/slack/).
+
