### 2021-03-04 3151a02
add remark about unicode in docstrings
