### 2021-03-05 FourierFlows/GeophysicalFlows.jl#203 @navidcy
Pfff... this is unfortunately due to the large size the gh-pages branch has reached...

You see, the way the binder urls work in Documenter.jl via Literate.jl (as far as I understand at least -- I may be wrong) is that it makes a clone of the documentation gh-pages repository, from which they get the `.toml` files to build the environment. Due to 1) the large number of examples we have incorporated into the Docs which include animations and 2) the several GeophysicalFlows.jl releases we made, the gh-pages branch size has become >1GB and thus binder errors!

I don't really know how to alleviate this problem. I've tried many times to clean up the gh-pages branch, asked around in slack and googled a lot... Still haven't figure how to deal with this in a simple manner. It gives me bad dreams... :)

Perhaps @mortenpi or @fredrikekre have some idea on how to make this work? If not, I'm tempted to simply remove the binder urls from the examples in the Documentation.
