### 2021-03-05 91f6faf
Update CONTRIBUTING.md
Co-authored-by: Hendrik Ranocha <ranocha@users.noreply.github.com>
