### 2021-03-05 3a5fa5e
decorate with Array() + remove binder links
