### 2021-03-05 b1b1a5d
all examples on GPU
