+++ examples/twodnavierstokes_stochasticforcing.jl
@@ -18,7 +18,7 @@ import GeophysicalFlows.TwoDNavierStokes: energy, enstrophy
 
 # ## Choosing a device: CPU or GPU
 
-dev = CPU()    # Device (CPU/GPU)
+dev = GPU()    # Device (CPU/GPU)
 nothing # hide
 
 
