+++ examples/multilayerqg_2layer.jl
@@ -16,7 +16,7 @@ import GeophysicalFlows.MultiLayerQG: energies
 
 # ## Choosing a device: CPU or GPU
 
-dev = CPU()     # Device (CPU/GPU)
+dev = GPU()     # Device (CPU/GPU)
 nothing # hide
 
 
