### 2021-03-05 #3053 @navidcy
@ranocha, I don't really see how I should reformat the file in view of your remark: "You should consider reformatting the license file LICENSE.md such that GitHub recognizes the license type automatically, cf. https://opensource.org/licenses/MIT"

Could you elaborate a bit perhaps?
