### 2021-03-05 FourierFlows/GeophysicalFlows.jl#203 @mortenpi
If the issue is the size of the `gh-pages` branch (which you can't really reduce, since you want the examples etc), I think the only way to make it work would be to store the notebooks in a separate repository. You'd need to customise `docs/make.jl` quite a bit to make that work though.
