### 2021-03-06 f15ac35
make all examples GPU friendly -- attempt 1
