### 2021-03-06 #3053 @navidcy
OK I see! Thanks.
Done via [1439a25](https://github.com/FourierFlows/GeophysicalFlows.jl/commit/1439a259e42556ae4d72f7d931ce8f0cdca07b0d).
