### 2021-03-06 4edb99f
update FourierFlows compat reqs
