+++ examples/Project.toml
@@ -7,5 +7,5 @@ Plots = "91a5bcdd-55d7-5caf-9e0b-520d859cae80"
 Printf = "de0858da-6303-5e67-8744-51eddeeeb8d7"
 
 [compat]
-FourierFlows = "≥ 0.6.10"
+FourierFlows = "≥ 0.6.12"
 Plots = "≥ 1.10.1"
