### 2021-03-06 994535d
adds warning when user wants MultiLayerQG on GPU
