### 2021-03-06 61b824f
some simplifications
