+++ examples/singlelayerqg_decaying_barotropic_equivalentbarotropic.jl
@@ -16,7 +16,7 @@ import GeophysicalFlows: peakedisotropicspectrum
 
 # ## Choosing a device: CPU or GPU
 
-dev = GPU()     # Device (CPU/GPU)
+dev = GPU()     # Devvice (CPU/GPU)
 nothing # hide
 
 

[…]
