### 2021-03-06 32dfa4f
final tweaks -- check on GPU now
