### 2021-03-06 1914a46
another try for GPU
