### 2021-03-07 17c5288
better parameter values
