### 2021-03-08 07405fe
add remark for ArrayType()
