### 2021-03-08 #3053 @navidcy
@ranocha, regarding your remarks on

- You should consider reformatting the license file LICENSE.md such that GitHub recognizes the license type automatically, cf. https://opensource.org/licenses/MIT
Done; see release [v0.11.5](https://github.com/FourierFlows/GeophysicalFlows.jl/releases/tag/v0.11.5)

- The currently latest release is v0.8.5 but the README tells people to cite v0.11.4 - that's really confusing and weird.
Done; see release [v0.11.5](https://github.com/FourierFlows/GeophysicalFlows.jl/releases/tag/v0.11.5)

- There is no explicit section "Statement of need" required for review. It would be good to structure the paper a bit more, e.g. by including sections like "Features" and "Related research and software" (although some related information is contained in the manuscript already).
- The first 3 paragraphs of the paper were aimed to summarize the statement of need. But we see your point, we'll rephrase/restructure to make it clearer to the reader!

- Would it be possible to leverage ApproxFun.jl (and/or related Matlab tools) for these problems?
This is something we've thought about. There is also some related discussion in https://github.com/FourierFlows/FourierFlows.jl/issues/243.
But this falls outside the scope of GeophysicalFlows.jl; it should be addressed/dealt in FourierFlows.jl

-  What is the reason for implementing the time stepping algorithms on your own instead of using something like OrdinaryDiffEq.jl?
Again, perhaps the quick short answer is "legacy; when we started FourierFlows.jl we didn't know that we could use DifferentialEquations.jl for timestepping". Later on we realized that as a possibility and it's in our mind. But again, that's outside of the scope of GeophysicalFlows.jl but rather something to be dear with within FourierFlows.jl

- How do you think about relations to Oceananigans.jl?
We totally forgot to remark on that in the paper! We'll do so. Several contributors are also heavily involved with Oceananigans.jl and, I guess, when we were thinking about relationship with other packages we didn't thought of Oceananigans.jl being some "other" package. :) We'll add a remark on the paper.

- From reading the manuscript and the README, it's not clear to me what kind of scalability to expect. Is it correct that this package is restricted to using a single CPU or GPU, mostly in two spatial dimensions? This should be made more clear.
You are correct, at the moment FourierFlows.jl is restricted to either a single GPU or a single CPU node (though multi-threading is possible and gives advantage). We'll add a remark in the README file and/or Docs.

- What about integration of functionality from PencilFFTs.jl for parallelism? Would something like this be possible?
Same as some of the remarks above: this is on the FourierFlows.jl to-do list and falls outside of the scope of GeophysicalFlows.jl

- I could not launch some examples in Binder as described in the documentation, see FourierFlows/GeophysicalFlows.jl#203
We removed binder links since the large size of the gh-pages branch rendered the binder unusable. See https://github.com/FourierFlows/GeophysicalFlows.jl/issues/203 and https://github.com/FourierFlows/GeophysicalFlows.jl/issues/207.

- I also tried to run the first few examples from the documentation on the GPU but got some errors like ERROR: scalar getindex is disallowed. It would be really nice to point the reader explicitly to some GPU examples or (even better) to make the CPU examples compatible with running them on the GPU.
Fixed with https://github.com/FourierFlows/GeophysicalFlows.jl/issues/203 and release [v0.11.5](https://github.com/FourierFlows/GeophysicalFlows.jl/releases/tag/v0.11.5). Now *all* examples run on GPU without complaints.
