### 2021-03-08 #3053 @ranocha
Nice work and good updates! Please find below two additional questions/comments.

- [ ] Could you please explain your reasons for creating several submodules such as `TwoDNavierStokes`? A more Julian approach seems to be to create some equations type and pass that around using multiple dispatch. That would probably make it easier to use GeophysicalFlows.jl as a library with multiple physical setups.
- [ ] What happens if I have multiple GPUs in one workstation and choose the device `GPU()`?
