### 2021-03-08 #3053 @pdebuyl
@navidcy the "Statement of need" is actually required for all JOSS papers since some time. It was not flagged for submission as the paper is not in the master branch for your repository, but you should add it. See https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain + the checklist. It should be section titled "Statement of need".
