### 2021-03-08 FourierFlows/GeophysicalFlows.jl#210 @navidcy
Add remark on multithreading/multiple GPU limitations that FourierFlows.jl imposes
We should add a remark in the README and in the Docs on this.

[This was mentioned by @ranocha's in their [review remarks](https://github.com/openjournals/joss-reviews/issues/3053#issuecomment-790762419).]
