### 2021-03-08 #3053 @navidcy
> @navidcy the "Statement of need" is actually required for all JOSS papers since some time. It was not flagged for submission as the paper is not in the master branch for your repository, but you should add it. See https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain + the checklist. It should be section titled "Statement of need".

Thanks for pointing this out -- I missed the fact that it was an actual requirement to have a section titles so. I looked at previous JOSS papers and followed their pattern (e.g. [Oceananigans.jl](https://joss.theoj.org/papers/10.21105/joss.02018)) and didn't see that. But sure, we'll restructure the paper to include a *Statement of need* section.
