### 2021-03-09 #3053 @navidcy
@glwagner, regarding @ranocha's question:
 > * [ ]  What happens if I have multiple GPUs in one workstation and choose the device `GPU()`?
do you have access to a machine with multiple GPUs to check what happens in this case?
