### 2021-03-09 #3053 @glwagner
I will check but I believe that `CUDA.jl` somehow generates a list of devices (possibly in a different order than the list returned by `nvidia-smi`) and selects the first device of the list. It's explained in [a blog post on the CUDA.jl website](https://juliagpu.org/2020-07-18-cuda_1.3/):

```julia
julia> using CUDA

shell> nvidia-smi
Mon Mar  8 20:39:35 2021
+-----------------------------------------------------------------------------+
| NVIDIA-SMI 440.33.01    Driver Version: 440.33.01    CUDA Version: 10.2     |
|-------------------------------+----------------------+----------------------+
| GPU  Name        Persistence-M| Bus-Id        Disp.A | Volatile Uncorr. ECC |
| Fan  Temp  Perf  Pwr:Usage/Cap|         Memory-Usage | GPU-Util  Compute M. |
|===============================+======================+======================|
|   0  TITAN V             Off  | 00000000:18:00.0 Off |                  N/A |
| 28%   34C    P8    23W / 250W |   5150MiB / 12066MiB |      0%      Default |
+-------------------------------+----------------------+----------------------+
|   1  TITAN V             Off  | 00000000:3B:00.0 Off |                  N/A |
| 34%   48C    P0    41W / 250W |     13MiB / 12066MiB |      0%      Default |
+-------------------------------+----------------------+----------------------+
|   2  TITAN V             Off  | 00000000:86:00.0 Off |                  N/A |
| 55%   75C    P2   138W / 250W |   5826MiB / 12066MiB |     98%      Default |
+-------------------------------+----------------------+----------------------+
|   3  TITAN V             Off  | 00000000:AF:00.0 Off |                  N/A |
| 28%   41C    P8    25W / 250W |     13MiB / 12066MiB |      0%      Default |
+-------------------------------+----------------------+----------------------+

+-----------------------------------------------------------------------------+
| Processes:                                                       GPU Memory |
|  GPU       PID   Type   Process name                             Usage      |
|=============================================================================|
|    0      1551      C   julia                                       3673MiB |
|    0     10610      G   /usr/libexec/Xorg                              9MiB |
|    0     10740      G   /usr/bin/gnome-shell                          10MiB |
|    0     41030      C   /home/ballen/julia-1.5.3/bin/julia          1443MiB |
|    2     36042      C   /home/sandre/julia-1.5.2/bin/julia          5813MiB |
+-----------------------------------------------------------------------------+

julia> device()
CuDevice(0): TITAN V

julia> collect(devices())
4-element Array{CuDevice,1}:
 CuDevice(0): TITAN V
 CuDevice(1): TITAN V
 CuDevice(2): TITAN V
 CuDevice(3): TITAN V

julia> device!(1)

julia> device()
CuDevice(1): TITAN V
```

We could implement a convenience interface to `device` by adding an argument to the `GPU()` constructor, something like

```julia
function GPU(device_tag = device())
    device!(device_tag)
    return GPU()
end
```

then users could write

```julia
dev = GPU(1)
```

which would select device 1 for the computation.

It's also possible to limit visible devices by setting the `CUDA_VISIBLE_DEVICES` environment variable prior to starting julia.
