### 2021-03-09 FourierFlows/GeophysicalFlows.jl#210 @glwagner
I think `CUDA.jl` may pick this GPU by default. I think the best solution is to link to `CUDA.jl` documentation for choosing a device. Users are also able to do fancier things, like run two problems side by side on different GPUs (some explanation is provided in the CUDA.jl docs for this).
