### 2021-03-09 FourierFlows/GeophysicalFlows.jl#210 @navidcy
Could you point to this explanation and I’ll add a not in our docs.
