### 2021-03-09 FourierFlows/GeophysicalFlows.jl#210 @glwagner
Here's some references:

* [Blog post on JuliaGPU about device-selection features added in CUDA 1.3](https://juliagpu.org/2020-07-18-cuda_1.3/). This is just a blog post so the syntax could go out of date.
* [Page in JuliaGPU docs about multi-GPU programming](https://juliagpu.gitlab.io/CUDA.jl/usage/multigpu/). This page is oriented towards people who want to use more than one GPU (not just select one of many).
* [Documentation of `CUDA.device`](https://juliagpu.gitlab.io/CUDA.jl/api/essentials/#CUDA.device). `device!` is adjacent. I don't see the function `devices` however.

Often the most straightforward approach to using mulitple GPUs is to launch the same script but with different `CUDA_VISIBLE_DEVICES` parameters. This approach is outside julia.

```
$ CUDA_VISIBLE_DEVICES=0 julia --project cool_script.jl
```

This launches julia but with only one device visible (device "0" from the output of `nvidia-smi`). This environment variable is described in nvidia's CUDA documentation:

https://developer.nvidia.com/blog/cuda-pro-tip-control-gpu-visibility-cuda_visible_devices/
