### 2021-03-10 #3053 @navidcy
@ranocha, regarding the items involving ApproxFun.jl, time stepping with DifferentialEquations.jl, and using PencilFFTs.jl:
We feel that there is nothing else to do after what we [replied above](https://github.com/openjournals/joss-reviews/issues/3053#issuecomment-792408616). But happy to discuss more on it if you think differently :)
