### 2021-03-11 #3053 @eviatarbach
I commend the authors for the excellent package. I just have a few more comments before I can recommend publication:

- I think you should also compare GeophysicalFlows.jl to [MAOOAM](https://gmd.copernicus.org/articles/9/2793/2016/) and [qgs](https://joss.theoj.org/papers/10.21105/joss.02597) in the paper, as they are also open-source packages with QG implementations.
- Please separate the paper into sections, namely a 'Statement of need' section and a 'State of the field' section (see the qgs paper for an example). Currently, everything is in the 'Summary' section.
- The installation instructions are not available on the online documentation, but only in the GitHub README.

Minor edit suggestions:
- L17: The comma should be a colon, semicolon, or period to avoid a run-on sentence.
- L29–30: "the Python's pyqg" -> "the Python package pyqg"
- L34: "is Python software" -> "is a Python package"
