### 2021-03-11 FourierFlows/GeophysicalFlows.jl#211 @navidcy
Add Installation Instructions section in Docs
We should add an Installation Instruction section in the Docs.

This was pointed out by @eviatarbach  in their [review](https://github.com/openjournals/joss-reviews/issues/3053#issuecomment-796390180).
