### 2021-03-11 6cd8694
adds installation instructions in docs
