### 2021-03-11 #3053 @navidcy
@eviatarbach, minor edits were taken care of by [054afc](https://github.com/FourierFlows/GeophysicalFlows.jl/commit/054afcdf59dbfb8ecd9d72f43a2524f727e1726a). Thanks.
