### 2021-03-11 4217d19
remove module from title + add links in examples
