+++ docs/src/modules/singlelayerqg.md
@@ -1,4 +1,4 @@
-# SingleLayerQG Module
+# SingleLayerQG
 
 ### Basic Equations
 
@@ -110,10 +110,18 @@ Other diagnostic include: [`energy_dissipation`](@ref GeophysicalFlows.SingleLay
 
 ## Examples
 
-- `examples/singlelayerqg_betadecay.jl`: A script that simulates decaying quasi-geostrophic flow on a beta-plane demonstrating zonation.
+- [`examples/singlelayerqg_betadecay.jl`](../generated/singlelayerqg_betadecay/): A script that 
+  simulates decaying quasi-geostrophic flow on a beta-plane demonstrating zonation.
 
-- `examples/singlelayerqg_betaforced.jl`: A script that simulates forced-dissipative quasi-geostrophic flow on a beta plane demonstrating zonation. The forcing is temporally delta-correlated with isotropic spatial structure with power in a narrow annulus in wavenumber space that corresponds to total wavenumber ``k_f``.
+- [`examples/singlelayerqg_betaforced.jl`](../generated/singlelayerqg_betaforced/): A script that 
+  simulates forced-dissipative quasi-geostrophic flow on a beta plane demonstrating zonation. 
+	The forcing is temporally delta-correlated with isotropic spatial structure with power in 
+	a narrow annulus in wavenumber space with total wavenumber ``k_f``.
 
-- `examples/singlelayerqg_decay_topography.jl`: A script that simulates two dimensional turbulence (barotropic quasi-geostrophic flow with ``\beta=0``) above topography.
+- [`examples/singlelayerqg_decay_topography.jl`](../generated/singlelayerqg_decay_topography/): 
+  A script that simulates two dimensional turbulence (barotropic quasi-geostrophic flow with 
+	``\beta=0``) above topography.
 
-- `examples/singlelayerqg_decaying_barotropic_equivalentbarotropic.jl`: A script that simulates two dimensional turbulence (``\beta=0``) with both infinite and finite Rossby radius of deformation and compares the evolution of the two.
+- [`examples/singlelayerqg_decaying_barotropic_equivalentbarotropic.jl`](../generated singlelayerqg_decaying_barotropic_equivalentbarotropic/): 
+  A script that simulates two dimensional turbulence (``\beta=0``) with both infinite and finite 
+	Rossby radius of deformation and compares the evolution of the two.
