+++ docs/src/modules/multilayerqg.md
@@ -1,4 +1,4 @@
-# MultiLayerQG Module
+# MultiLayerQG
 
 ### Basic Equations
 
@@ -146,4 +146,5 @@ GeophysicalFlows.MultiLayerQG.fluxes
 
 ## Examples
 
- - `examples/multilayerqg_2layer.jl`: A script that simulates the growth and equilibration of baroclinic eddy turbulence in the Phillips 2-layer model.
+ - [`examples/multilayerqg_2layer.jl`](../generated/multilayerqg_2layer/): A script that 
+   simulates the growth and equilibration of baroclinic eddy turbulence in the Phillips 2-layer model.
