### 2021-03-11 81fc154
appveyor badge + rephrase installing on Julia v1.0.5
