+++ docs/src/stochastic_forcing.md
@@ -64,14 +64,14 @@ popular ways for computing such stochastic integrals are:
 
 ```math
 \begin{aligned}
-{\color{Green} \text{Itô}}&: {\color{Green}\int_{t_0}^{t} g(x(s)) \, \mathrm{d} W_s \approx \sum_{j} g \left ( x(t_j) \right )(W_{j+1} - W_j)} , \\
-{\color{Magenta} \text{Stratonovich}}&: {\color{Magenta} \int_{t_0}^{t} g(x(s)) \, \mathrm{d} W_s \approx \sum_{j} g \left (x \left (\tfrac{1}{2}(t_j + t_{j+1}) \right ) \right)(W_{j+1} - W_j)} .
+{\color{Green} \text{Itô}} &: {\color{Green}\int_{t_0}^{t} g(x(s)) \, \mathrm{d} W_s \approx \sum_{j} g \left ( x(t_j) \right )(W_{j+1} - W_j)} , \\
+{\color{Magenta} \text{Stratonovich}} &: {\color{Magenta} \int_{t_0}^{t} g(x(s)) \, \mathrm{d} W_s \approx \sum_{j} g \left (x \left (\tfrac{1}{2}(t_j + t_{j+1}) \right ) \right)(W_{j+1} - W_j)} .
 \end{aligned}
 ```
 
 The difference in the two calculi above lies in the point at which we choose to evaluate ``g(x)``:
-we take the start of the time-interval for ${\color{Green} \text{Itô, } t_j}$, while we use
-the mid-point for ${\color{Magenta}\text{Stratonovich, } \tfrac{1}{2}(t_j+t_{j+1})}$. In the case the 
+we take the start of the time-interval for ``{\color{Green} \text{Itô, } t_j}``, while we use
+the mid-point for ``{\color{Magenta}\text{Stratonovich, } \tfrac{1}{2}(t_j+t_{j+1})}``. In the case the 
 stochastic noise is additive, i.e., its prefactor ``g`` does not depend on the state ``x_t``,
 then the two interpretations coincide. When the noise does depend on the state of the system, 
 i.e., ``g=g(x(t))``, then the two interpretations above give thoroughly different results. This
@@ -89,8 +89,8 @@ differential form as:
 
 ```math
 \begin{aligned}
-{\color{Green}\text{Itô}}&: {\color{Green}\mathrm{d} x_t = f(x_t) \mathrm{d} t + g(x_t) \mathrm{d} W_t} , \\
-{\color{Magenta}\text{Stratonovich}}&: {\color{Magenta}\mathrm{d} x_t = f(x_t) \mathrm{d} t + g(x_t) \circ \mathrm{d} W_t} .
+{\color{Green}\text{Itô}} &: {\color{Green}\mathrm{d} x_t = f(x_t) \mathrm{d} t + g(x_t) \mathrm{d} W_t} , \\
+{\color{Magenta}\text{Stratonovich}} &: {\color{Magenta}\mathrm{d} x_t = f(x_t) \mathrm{d} t + g(x_t) \circ \mathrm{d} W_t} .
 \end{aligned}
 ```
 
@@ -103,8 +103,8 @@ rule must be used:
 
 ```math
 \begin{aligned}
[…]
