### 2021-03-11 bd1589a
change filename decay->decaying + fix movie names
