### 2021-03-15 #3053 @navidcy
@pdebuyl I plan to respond to everything and push a new version of the paper ideally in the next 2 weeks. Does this sound reasonable?
