+++ paper/paper.bib
@@ -1,3 +1,33 @@
+@article{Oceananigans,
+  title={Oceananigans.jl: {Fast} and friendly geophysical fluid dynamics on {GPUs}},
+  author={Ramadhan, Ali and Wagner, Gregory LeClaire and Hill, Chris and Campin, Jean-Michel and Churavy, Valentin and Besard, Tim and Souza, Andre and Edelman, Alan and Ferrari, Raffaele and Marshall, John},
+  journal={Journal of Open Source Software},
+  volume={5},
+  number={53},
+  pages={2018},
+  year={2020}
+}
+
+@article{MAOOAM,
+	author = {De Cruz, L. and Demaeyer, J. and Vannitsem, S.},
+	doi = {10.5194/gmd-9-2793-2016},
+	journal = {Geoscientific Model Development},
+	volume = {9},
+	number = {8},
+	pages = {2793--2808},
+	title = {The Modular Arbitrary-Order Ocean-Atmosphere Model: \textsc{maooam}~v1.0},
+	year = {2016}}
+
+@article{qgs,
+	author = {Demaeyer, J. and De Cruz, L. and Vannitsem, S.},
+	doi = {10.21105/joss.02597},
+	journal = {Geoscientific Model Development},
+	volume = {5},
+	number = {56},
+	pages = {2597},
+	title = {qgs: {A} flexible {Python} framework of reduced-order multiscale climate models},
+	year = {2020}}
+
 @software{FourierFlows,
   author       = {Constantinou, N. C. and
                   Wagner, L. C. and
