### 2021-03-30 7db7d38
adds remarks about MAOOAM and qgs
