+++ paper/paper.bib
@@ -90,7 +90,7 @@
   title = {Advective structure functions in anisotropic two-dimensional turbulence},
   author = {Pearson, B. C. and Pearson, J. L. and Fox-Kemper, B.},
   journal = {J. Fluid Mech.},
-  year = {Under Review}
+  year = {In Press}
 }
 
 @article{Karrasch2020,
