### 2021-04-02 19ada9e
adds most of @BrodiePearson's remarks
