+++ paper/paper.md
@@ -49,7 +49,7 @@ pseudospectral numerical methods and leverage the framework provided by the `Fou
 
 # Statement of need
 
-Conceptual models in simple domains often provide stepping stones for better understanding in geophysical and astrophysical settings: atmospheric, oceanic, climate, and planetary atmospheres. These conceptual models are used in research but also are of great value for helping students in class to grasp on new concepts and phenomena. Oftentimes people end up coding their own versions of solvers for the same partial differential equation.
+Conceptual models in simple domains often provide stepping stones for better understanding geophysical and astrophysical systems, particularly the atmospheres and oceans of Earth and other planets. These conceptual models are used in research but also are of great value for helping students in class to grasp on new concepts and phenomena. Oftentimes people end up coding their own versions of solvers for the same partial differential equations for research or classwork. `GeophysicalFlows.jl` package is designed to be easily utilized and adaptable for a wide variety of both research and pedagogical purposes.
 
 On top of the above-mentioned needs, the recent explosion of machine-learning applications in atmospheric and oceanic sciences advocates for the need that solvers for partial differential equations can be run on GPUs. 
 
@@ -81,8 +81,8 @@ the Python package `pyqg` [@pyqg]. Beyond their base language, the major differe
 packages are that `GeophysicalFlows.jl` can be run on GPUs or CPUs and leverages a separate 
 package (`FourierFlows.jl`; which is continuously developed) to solve differential equations 
 and compute diagnostics, while `pyqg` can only be run on CPUs and uses a self-contained kernel. 
-Dedalus [@Burns2020] is Python package with an intuitive script-based interface that uses spectral 
-methods to solve general partial differential equations, such as the ones within `GeophysicalFlows.jl`. `Oceananigans.jl` [@Oceananigans] is not bad either. It also runs on GPUs, it allows for more variety of boundary conditions but is slower and does not have spectral accuracy. (Greg could you write a blurp (e.g. 1 sentence) comparing it with `GeophysicalFlows.jl`? Just delete my sentence all together if you like --- I just wrote up something simply as a placeholder.) The `MAOOAM` [@MAOOAM] package (and its Python implementation `qgs` [@qgs]) provides a reduced-order quasi-geostrophic coupled ocean–atmosphere model. These packages are fully spectral and thus allow to study how the solutions depend on the number of spectral-mode truncation in the ocean and atmosphere. The physics in `MAOOAM` and `qgs` share similarities with the `SingleLayerQG` and `MultiLayerQG` modules of `GeophysicalFlows.jl`; but neither `MAOOAM` nor `qgs` can run on GPUs. There are also some other isolated codes/scripts in personal websites and in open-source public repositories that have similar functionality as some `GeophysicalFlows.jl` modules. 
+Dedalus [@Burns2020] is a Python package with an intuitive script-based interface that uses spectral 
[…]
