+++ paper/paper.md
@@ -82,7 +82,7 @@ packages are that `GeophysicalFlows.jl` can be run on GPUs or CPUs and leverages
 package (`FourierFlows.jl`; which is continuously developed) to solve differential equations 
 and compute diagnostics, while `pyqg` can only be run on CPUs and uses a self-contained kernel. 
 Dedalus [@Burns2020] is a Python package with an intuitive script-based interface that uses spectral 
-methods to solve general partial differential equations, such as the ones within `GeophysicalFlows.jl`. `Oceananigans.jl` [@Oceananigans] is an incompressible fluid solver that can be run in 1-D, 2-D, and 3-D domains. Both Oceananigans.jl and GeophysicalFlows.jl contain modules for two-dimensional flow, but the other examples provided with Oceananigans focus on three-dimensional flows which are generally present at smaller scales than the dynamics of GeophysicalFlows.jl's other modules. The `MAOOAM` [@MAOOAM] package, and its expanded Python implementation `qgs` [@qgs], simulate two atmospheric layers with QG dynamics, above either land or an oceanic layer with reduced-gravity QG dynamics. The dynamics of individual layers have overlap with the `MultiLayerQG` and `SingleLayerQG` modules, however the layer configuration of `MOAAM` and `qgs` is specifically designed to study the dynamics of Earth's mid-latitude atmosphere. Neither `MAOOAM` nor `qgs` can run on GPUs. There exist also some other isolated codes/scripts in personal websites and in open-source public repositories that have similar functionality as some `GeophysicalFlows.jl` modules. 
+methods to solve general partial differential equations, such as the ones within `GeophysicalFlows.jl`. `Oceananigans.jl` [@Oceananigans] is a fluid solver focussed on the Navier-Stokes equations under the Boussinesq approximation. `Oceananigans.jl` also runs on GPUs, and it allows for more variety of boundary conditions but it does not have spectral accuracy as it uses finite-volume discretization methods. The `MAOOAM` [@MAOOAM] package, and its expanded Python implementation `qgs` [@qgs], simulate two atmospheric layers with QG dynamics, above either land or an oceanic fluid layer with reduced-gravity QG dynamics. The dynamics of individual layers have overlap with the `MultiLayerQG` and `SingleLayerQG` modules, however the layer configuration of `MOAAM` and `qgs` is specifically designed to study the dynamics of Earth's mid-latitude atmosphere. Neither `MAOOAM` nor `qgs` can run on GPUs. Finally, there exist also some other isolated codes/scripts in personal websites and in open-source public repositories that have similar functionality as some `GeophysicalFlows.jl` modules but usually come without any or poor documentation and typically are not tested.
 
 `GeophysicalFlows.jl` can be used to investigate a variety of scientific research questions 
 thanks to its various modules and high customizability, and its ease-of-use makes it an ideal 
