+++ paper/paper.bib
@@ -21,11 +21,11 @@
 @article{qgs,
 	author = {Demaeyer, J. and De Cruz, L. and Vannitsem, S.},
 	doi = {10.21105/joss.02597},
-	journal = {Geoscientific Model Development},
+	journal = {Journal of Open Source Software},
 	volume = {5},
 	number = {56},
 	pages = {2597},
-	title = {qgs: {A} flexible {Python} framework of reduced-order multiscale climate models},
+	title = {{qgs}: {A} flexible {Python} framework of reduced-order multiscale climate models},
 	year = {2020}}
 
 @software{FourierFlows,
