### 2021-04-05 2a66ca9
address everyones comments + attempt to bullet-point comparisons
