+++ paper/paper.bib
@@ -113,6 +113,18 @@
   url = {https://github.com/navidcy/KolmogorovFlow}
 }
 
+@misc{QG_tracer_advection,
+  author = {Bisits, J. and Constantinou, N. C.},
+  title = {QG_tracer_advection},
+  year = {2021},
+  publisher = {GitHub},
+  journal = {GitHub repository},
+  url = {https://github.com/jbisits/QG_tracer_advection}
+}
+
+
+https://github.com/jbisits/QG_tracer_advection
+
 @misc{GeophysicalFlows-Examples,
   author = {Constantinou, N. C. and Wagner, G. L.},
   title = {GeophysicalFlows-Examples},
