### 2021-04-05 7c4fe0e
bullet points v2
