### 2021-04-05 811eb25
more GPU-related remarks
