+++ src/utils.jl
@@ -12,8 +12,8 @@ function lambdipole(U, R, grid::TwoDGrid{T, A}; center=(mean(grid.x), mean(grid.
   x, y = gridpoints(grid)
   xc, yc = center
   r = @. sqrt( (x - xc)^2 + (y - yc)^2 )
-  CUDA.@allowscalar besselj1 = A([besselj(1, k * r[i, j]) for i=1:grid.nx, j=1:grid.ny])
-  q = @. q₀ * besselj1 * (y - yc) / r
+  besseljorder1 = CUDA.@allowscalar A([besselj(1, k * r[i, j]) for i=1:grid.nx, j=1:grid.ny])
+  q = @. q₀ * besseljorder1 * (y - yc) / r
   @. q = ifelse(r >= R, 0, q)
   
   return q
