### 2021-04-06 3c02b3a
fix k² in calcS⁻¹
