### 2021-04-06 fb7a002
bumps version to 0.11.6
