### 2021-04-06 #3053 @navidcy
#### Response to @ranocha's remaining remarks

- There is no explicit section "Statement of need" required for review. It would be good to structure the paper a bit more, e.g. by including sections like "Features" and "Related research and software" (although some related information is contained in the manuscript already).
- How do you think about relations to Oceananigans.jl?

The revised manuscript now addresses both of the above remarks.

- From reading the manuscript and the README, it's not clear to me what kind of scalability to expect. Is it correct that this package is restricted to using a single CPU or GPU, mostly in two spatial dimensions? This should be made more clear.

-  What happens if I have multiple GPUs in one workstation and choose the device GPU()?

Yeap, that's correct. FourierFlows.jl is restricted to single CPU or GPU. Multi-threading can be used to speed up ffts.

We have added a section in the package's [README](https://github.com/FourierFlows/GeophysicalFlows.jl#readme) as well as in the documentation (v0.11.6 and later) as well as remarks in [FourierFlows.jl README](https://github.com/FourierFlows/FourierFlows.jl#readme) and [Docs](https://fourierflows.github.io/FourierFlowsDocumentation/stable/gpu/)

- Could you please explain your reasons for creating several submodules such as TwoDNavierStokes? A more Julian approach seems to be to create some equations type and pass that around using multiple dispatch. That would probably make it easier to use GeophysicalFlows.jl as a library with multiple physical setups.

Indeed, some of the modules might have overlapping functionality. For example, the equations that  `TwoDNavierStokes` module solves is reproduced by `SingleLayerQG` when setting `β=0`, `\ell = 0` and no topography. The reason we have done so was party because different disciplines feel more comfortable, i.e., mathematicians working with the 2D Navier-Stokes or 2D Euler equations don't feel quite "at home" when presented with the "single layer quasi-geostrophic equations". We see the reviewer's point thought that we could have some common functionality under the hood and then the various modules use multiple dispatch to use and/or enhance that functionality. This, though, requires a big restructure of the package and we decided to postpone it for future. Happy to discuss more on this.

#### Response to @eviatarbach's remaining remarks

- I think you should also compare GeophysicalFlows.jl to MAOOAM and qgs in the paper, as they are also open-source packages with QG implementations.
-
- Please separate the paper into sections, namely a 'Statement of need' section and a 'State of the field' section (see the qgs paper for an example). Currently, everything is in the 'Summary' section.

The revised paper addresses both of these remarks; see section "Stat of the field" in the revised version of the paper.
