### 2021-04-06 #3053 @eviatarbach
Thank you for addressing my points! I am happy to recommend publication as is.

Note that MAOOAM has Python and Lua implementations in addition to the Fortran.
