### 2021-04-07 d980973
add julia ref + update figure
