+++ paper/paper.bib
@@ -1,3 +1,14 @@
+@article{Bezanson2017,
+  title   = {Julia: {A} fresh approach to numerical computing},
+  author  = {Bezanson, Jeff and Edelman, Alan and Karpinski, Stefan and Shah, Viral B.},
+  journal = {SIAM Review},
+  volume  = {59},
+  number  = {1},
+  pages   = {65--98},
+  year    = {2017},
+  doi     = {10.1137/141000671}
+}
+
 @article{Oceananigans,
   title={Oceananigans.jl: {Fast} and friendly geophysical fluid dynamics on {GPUs}},
   author={Ramadhan, Ali and Wagner, Gregory LeClaire and Hill, Chris and Campin, Jean-Michel and Churavy, Valentin and Besard, Tim and Souza, Andre and Edelman, Alan and Ferrari, Raffaele and Marshall, John},
