+++ paper/paper.bib
@@ -26,7 +26,7 @@
 	volume = {9},
 	number = {8},
 	pages = {2793--2808},
-	title = {The Modular Arbitrary-Order Ocean-Atmosphere Model: \textsc{maooam}~v1.0},
+	title = {The modular arbitrary-order ocean-atmosphere model: \textsc{maooam}~v1.0},
 	year = {2016}}
 
 @article{qgs,
@@ -43,11 +43,10 @@
   author       = {Constantinou, N. C. and
                   Wagner, L. C. and
                   Palóczy, A.},
-  title        = {FourierFlows/FourierFlows.jl: v0.6.10},
-  month        = jan,
-  year         = 2021,
+  title        = {FourierFlows/FourierFlows.jl: v0.6.13},
+  year         = {2021},
   publisher    = {Zenodo},
-  version      = {v0.6.10},
+  version      = {v0.6.13},
   doi          = {10.5281/zenodo.4460734},
   url          = {https://doi.org/10.5281/zenodo.4460734}
 }
@@ -64,9 +63,9 @@
 }
 
 @article{Burns2020,
-  title = {Dedalus: A flexible framework for numerical simulations with spectral methods},
+  title = {Dedalus: {A} flexible framework for numerical simulations with spectral methods},
   author = {Burns, K. J. and Vasil, G. M. and Oishi, J. S. and Lecoanet, D. and Brown, B. P.},
-  journal = {Phys. Rev. Research},
+  journal = {Physical Review Research},
   volume = {2},
   issue = {2},
   pages = {023068},
@@ -88,9 +87,8 @@
                   Penn, J. and
                   Wolfe, C. L. P. and
                   Boas, B. V.},
-  title        = {pyqg/pyqg: v0.3.0},
-  month        = nov,
-  year         = 2019,
+  title        = {{pyqg/pyqg}: v0.3.0},
+  year         = {2019},
   publisher    = {Zenodo},
   version      = {v0.3.0},
   doi          = {10.5281/zenodo.3551326},
@@ -133,9 +131,6 @@
   url = {https://github.com/jbisits/QG_tracer_advection}
 }
 
-
-https://github.com/jbisits/QG_tracer_advection
-
 @misc{GeophysicalFlows-Examples,
   author = {Constantinou, N. C. and Wagner, G. L.},
   title = {GeophysicalFlows-Examples},
@@ -153,4 +148,3 @@ https://github.com/jbisits/QG_tracer_advection
   journal = {GitHub repository},
   url = {https://github.com/navidcy/CLExWinterSchool2020}
 }
-
