### 2021-04-07 de3d641
attempt to test both in Julia v1.5 and v1.6
