### 2021-04-08 aa1acb1
let's see if CUDA v2.4.2 allows tests on GPU
