### 2021-04-10 70f1818
CompatHelper: bump compat for "CUDA" to "3.0"
