### 2021-04-12 c6df4d0
some typos + fixes
