+++ docs/src/contributing.md
@@ -26,7 +26,7 @@ New contributors should make sure to read that guide.
   periodic boundary conditions.
 
 If you're interested in working on something, let us know by commenting on
-existing issues or by opening a new issue if. This is to make sure no one else
+existing issues or by opening a new issue. This is to make sure no one else
 is working on the same issue and so we can help and guide you in case there
 is anything you need to know beforehand.
 
@@ -76,7 +76,7 @@ or by requesting something you think is missing.
   git remote add geophysicalflows https://github.com/FourierFlows/GeophysicalFlows.jl.git
   ```
 * Create the development environment by opening Julia via `julia --project` then
-  typing in `] instantiate`. This will install all the dependencies in the Project.toml
+  typing in `] instantiate`. This will install all the dependencies in the `Project.toml`
   file.
 * You can test to make sure GeophysicalFlows.jl works by typing in `] test` which will run all
   the tests (this can take a while). In an ideal world you should run the tests on a machine
@@ -87,7 +87,7 @@ Your development environment is now ready!
 
 ## Pull Requests
 
-Changes and contributions should be made via GitHub pull requests against the ``master`` branch.
+Changes and contributions should be made via GitHub pull requests against the `master` branch.
 
 When you're done making changes, commit the changes you made. Chris Beams has written 
 a [guide](https://chris.beams.io/posts/git-commit/) on how to write good commit messages.
@@ -115,7 +115,7 @@ You can preview how the Documentation will look like after merging by building t
 locally. To do that, from the main directory of your local repository call
 
 ```
-julia --project=docs/ -e 'using Pkg; Pkg.instantiate(); Pkg.develop(PackageSpec(path=pwd()))'
+julia --project=docs/ -e 'using Pkg; Pkg.develop(PackageSpec(path=pwd())); Pkg.instantiate()'
 julia --project=docs/ docs/make.jl
 ```
  
@@ -123,4 +123,5 @@ and then open `docs/build/index.html` in your favorite browser.
 
 ## Credits
 
-This contributor's guide is heavily based on the excellent [MetPy contributor's guide](https://github.com/Unidata/MetPy/blob/master/CONTRIBUTING.md) and on its "cover" made by [Oceananigans.jl](https://clima.github.io/OceananigansDocumentation/stable/contributing/).
+This contributor's guide is heavily based on the [MetPy contributor's guide](https://github.com/Unidata/MetPy/blob/master/CONTRIBUTING.md) 
+and on its "cover" made by [Oceananigans.jl](https://clima.github.io/OceananigansDocumentation/stable/contributing/).
