### 2021-04-12 4bfee3c
optimize stochastic-forcing calcF
