### 2021-04-12 63789cb
add CUDA in docs/Project.toml + in some examples
