### 2021-04-12 #3053 @pdebuyl
Thank you @ranocha and @eviatarbach for the review!

@navidcy I will now take over for the next step of the editorial procedure https://joss.readthedocs.io/en/latest/editing.html#after-reviewers-recommend-acceptance . When I am done (after possibly asking minor changes), I'll notify the eic for publication.
