### 2021-04-12 #3053 @glwagner
> > Could you please explain your reasons for creating several submodules such as TwoDNavierStokes? A more Julian approach seems to be to create some equations type and pass that around using multiple dispatch. That would probably make it easier to use GeophysicalFlows.jl as a library with multiple physical setups.

> Indeed, some of the modules might have overlapping functionality. For example, the equations that TwoDNavierStokes module solves is reproduced by SingleLayerQG when setting β=0, \ell = 0 and no topography. The reason we have done so was party because different disciplines feel more comfortable, i.e., mathematicians working with the 2D Navier-Stokes or 2D Euler equations don't feel quite "at home" when presented with the "single layer quasi-geostrophic equations". We see the reviewer's point thought that we could have some common functionality under the hood and then the various modules use multiple dispatch to use and/or enhance that functionality. This, though, requires a big restructure of the package and we decided to postpone it for future. Happy to discuss more on this.

I apologize for the late response --- I just wanted to add to @navidcy's response!

I think it's a valid criticism that the design of `GeophysicalFlows.jl` / `FourierFlows.jl` isn't built around multiple dispatch but instead relies on passing functions around (in particular, `FourierFlows.Equation` stores the function `calcN!` as data). I'd argue, however, that whether or not the code is organized into modules isn't the main issue. Rather than issue is how PDE solvers are implemented in the `FourierFlows.jl` framework.

To explain the two possibilities we need to distinguish from "module writers" (who are users of `FourierFlows.jl`, like the writers of `GeophysicalFlows.jl` modules) and "users" (who use `GeophysicalFlows.jl`").

The design `FourierFlows.jl` currently implements is that it defined a fundamental type `Problem`. Module writers then implement a PDE by building instances of `Problem`; the crucial function `calcN!` is stored in `Problem.eqn.calcN!`. In this design, the PDE being solved can only be inferred rather circuitously from the type of `prob = Problem(...)` (for example, if `typeof(prob.eqn.calcN!) === typeof(GeophysicalFlows.TwoDimNavierStokes)`, then we know that `prob` solves the two-dimensional Navier-Stokes equations). There is only one `stepforward!` function for `FourierFlows.Problem`. I think its a valid criticism that this design is not "ideally Julian".

An alternative design would define a fundamental abstract type (such as `AbstractProblem`) and then ask module writers to implement a new problem type. In this design `FourierFlows.jl` would implement a PDE "template" by defining placeholder functions (such as `calculate_nonlinear_term!(::ProblemType, args...)`, which module writers would extend for their PDE.

I don't think we have a great argument for taking the first approach over the second other than the knowledge of the developers at the time the code was written. At the end of the day I don't think that the first approach we currently use is _that_ limiting even though it is imperfect julia design. Because our imperfect design hasn't proved to be much of a barrier to implementing new PDEs (the difficult part is always writing `calcN!`, which is essentially the same task in either design for module writers), we haven't been motivated to undertake a redesign of `FourierFlows.jl` yet.

But someday, it might be worth it.
