### 2021-04-12 ac9d748
few more typos and rephrases
