### 2021-04-13 fdede9f
hopefully calcF! work as expected now
