### 2021-04-13 9874f71
finalize and check all examples on GPU
