+++ examples/barotropicqgql_betaforced.jl
@@ -6,11 +6,10 @@
 # a beta plane under the *quasi-linear approximation*. The dynamics include 
 # linear drag and stochastic excitation.
 
-using FourierFlows, Plots, Statistics, Printf, Random, CUDA
+using CUDA, FourierFlows, Random, Statistics, Printf, Plots
 
 using FourierFlows: parsevalsum
 using FFTW: irfft
-using Random: seed!
 using Statistics: mean
 
 import GeophysicalFlows.BarotropicQGQL
@@ -19,7 +18,7 @@ import GeophysicalFlows.BarotropicQGQL: energy, enstrophy
 
 # ## Choosing a device: CPU or GPU
 
-dev = CPU()     # Device (CPU/GPU)
+dev = GPU()     # Device (CPU/GPU)
 nothing # hide
 
 
@@ -46,12 +45,12 @@ nothing # hide
 # We force the vorticity equation with stochastic excitation that is delta-correlated in time 
 # and while spatially homogeneously and isotropically correlated. The forcing has a spectrum 
 # with power in a ring in wavenumber space of radius ``k_f`` (`forcing_wavenumber`) and width 
-# ``\delta k_f`` (`forcing_bandwidth`), and it injects energy per unit area and per unit time 
+# ``δ_f`` (`forcing_bandwidth`), and it injects energy per unit area and per unit time 
 # equal to ``\varepsilon``. That is, the forcing covariance spectrum is proportional to 
-# ``\exp{[-(|\bm{k}| - k_f)^2 / (2 \delta k_f^2)]}``.
+# ``\exp{[-(|\bm{k}| - k_f)^2 / (2 δ_f^2)]}``.
 
-forcing_wavenumber = 14.0 * 2π/L  # the central forcing wavenumber for a spectrum that is a ring in wavenumber space
-forcing_bandwidth  = 1.5  * 2π/L  # the width of the forcing spectrum 
+forcing_wavenumber = 14.0 * 2π/L  # the forcing wavenumber, `k_f`, for a spectrum that is a ring in wavenumber space
+forcing_bandwidth  = 1.5  * 2π/L  # the width of the forcing spectrum, `δ_f`
 ε = 0.001                         # energy input rate by the forcing
 
 grid = TwoDGrid(dev, n, L)
@@ -61,9 +60,13 @@ K = @. sqrt(grid.Krsq)            # a 2D array with the total wavenumber
 forcing_spectrum = @. exp(-(K - forcing_wavenumber)^2 / (2 * forcing_bandwidth^2))
 ε0 = parsevalsum(forcing_spectrum .* grid.invKrsq / 2, grid) / (grid.Lx * grid.Ly)
 @. forcing_spectrum *= ε/ε0       # normalize forcing to inject energy at rate ε
-
-seed!(1234) # reset of the random number generator for reproducibility
 nothing # hide
+
+
+# We reset of the random number generator for reproducibility
+if dev==CPU(); Random.seed!(1234); else; CUDA.seed!(1234); end
+nothing # hide
+
 
 # Next we construct function `calcF!` that computes a forcing realization every timestep.
 # First we make sure that if `dev=GPU()`, then `CUDA.rand()` function is called for random
