### 2021-04-13 b2ee085
adds CUDA in examples/Project.toml
