+++ examples/Project.toml
@@ -10,5 +10,5 @@ Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
 Statistics = "10745b16-79ce-11e8-11f9-7d13ad32a3b2"
 
 [compat]
-FourierFlows = "≥ 0.6.12"
+FourierFlows = "≥ 0.6.16"
 Plots = "≥ 1.10.1"
