+++ Project.toml
@@ -21,7 +21,7 @@ Statistics = "10745b16-79ce-11e8-11f9-7d13ad32a3b2"
 CUDA = "^1, ^2.4.2"
 DocStringExtensions = "^0.8"
 FFTW = "^1"
-FourierFlows = "^0.6.12"
+FourierFlows = "^0.6.16"
 JLD2 = "^0.1, ^0.2, ^0.3, ^0.4"
 Reexport = "^0.2, ^1"
 SpecialFunctions = "^0.10, ^1"
