### 2021-04-13 #3053 @pdebuyl
Thank you @glwagner for the extra information. I understand the appeal for being "pythonic", "julian", etc, but indeed software authors can still make architectural choices that deviate from that :-)
