### 2021-04-13 #3053 @pdebuyl
@navidcy I can't guess what to do with the very sparse installation instructions. `] add GeophysicalFlows` does not help much, can you let me know what to do. (and update the readme and also explain installation in the web documentation?)
