### 2021-04-13 #3053 @navidcy
> @navidcy I can't guess what to do with the very sparse installation instructions. `] add GeophysicalFlows` does not help much, can you let me know what to do. (and update the readme and also explain installation in the web documentation?)

@pdebuyl the docs include much more detail on how to install.
https://fourierflows.github.io/GeophysicalFlowsDocumentation/dev/installation_instructions/
Could you please let me know if indeed the doc instructions seem clear to you? If so, I can update the readme to reflect what the docs include. If you still feel docs instructions are not clear then I’ll update both.
