### 2021-04-13 #3053 @pdebuyl
Thanks for the reply. I was not familiar with the `]` syntax for accessing Pkg. It does not work by copy-paste, is that normal?

I must have had a outdated cache for the documentation, the current version is fine, thanks. The readme version is then the one that didn't allow me to proceed.

I asked "add GeophysicalFlows" but still had to install explicitly FourierFlows, FFTW, and Plot. That should be clear from the doc as well.
