### 2021-04-14 #3053 @navidcy
OK, @pdebuyl, I see... you probably tried to run the examples and run into that because, e.g., `Plots.jl` is not a dependency of `GeophysicalFlows.jl` so it must be installed separately.

We'll add a remark on that in the examples. And we'll try to clean up/smooth what's been re-exported by which package so, e.g., all `FourierFlows.jl` namespace is available to the user as soon they call `using GeophysicalFlows`.

I'll ping you here when these changes are made.
