### 2021-04-14 33c5385
reduce using Packages in examples + add comment for installing deps
