+++ docs/make.jl
@@ -1,5 +1,3 @@
-push!(LOAD_PATH, joinpath(@__DIR__, "..")) # add GeophysicalFlows to environment stack
-
 using
   Documenter,
   Literate,
