+++ docs/make.jl
@@ -1,7 +1,7 @@
 using
   Documenter,
   Literate,
-  Plots,  # to not capture precompilation output
+  Plots,   # to not capture precompilation output
   GeophysicalFlows
 
 # Gotta set this environment variable when using the GR run-time on Travis CI.
