### 2021-04-14 #3053 @pdebuyl
Thank you @navidcy for the detailed reply. Regarding the changes that you propose, my feedback is the one of a user testing the first example. I don't know if this warrants changing the way FourierFlows and GeophysicalFlows should expose the API of imported packages. The note about installing Plots.jl is of course welcome in the example section or in the installation section. For the rest, it is important that FFTW and FourierFlows are installed as dependencies.
