### 2021-04-14 #3053 @navidcy
@pdebuyl, I see your point.

But we discussed with @glwagner and thought that this way would feel much more natural to the `GeophysicalFlows.jl` users. The PR https://github.com/FourierFlows/GeophysicalFlows.jl/pull/229 adds remarks in all examples for installing packages that are not dependency of GeophysicalFlows.jl (e.g., Plots.jl) and also simplifies the packages needed to be called at the beginning by the users. I'll wait for a review before I merge. I'll ping you when it's ready for you to have a look.
