### 2021-04-15 4ed2f63
update installation instructions in readme
