### 2021-04-15 #3053 @navidcy
@pdebuyl, we've tagged a new release GeophysicalFlows.jl v0.12.0 which includes:
- updated installation instructions in Readme
- simpler loading of packages in examples (e.g., users don't need to call `using FourierFlows` nor `using CUDA` any more)
- remarks on the extra packages needed to be installed for each experiment

Have a look and let me know how that looks. Note that if you've installed GeophysicalFlows.jl already you might need to update to the latest version (something that users can do following instructions [in the docs](https://fourierflows.github.io/GeophysicalFlowsDocumentation/stable/installation_instructions/)).
