### 2021-04-15 #3053 @pdebuyl
I have the error

```
julia> ζ₀ = peakedisotropicspectrum(grid, k₀, E₀, mask=prob.timestepper.filter)
ERROR: UndefVarError: irfft not defined
Stacktrace:
 [1] peakedisotropicspectrum(::TwoDGrid{Float64,Array{Float64,2},StepRangeLen{Float64,Base.TwicePrecision{Float64},Base.TwicePrecision{Float64}},FFTW.cFFTWPlan{Complex{Float64},-1,false,2,UnitRange{Int64}},FFTW.rFFTWPlan{Float64,-1,false,2,UnitRange{Int64}}}, ::Int64, ::Float64; mask::Array{Float64,2}, allones::Bool) at /home/pierre/.julia/packages/GeophysicalFlows/06GdS/src/utils.jl:47
 [2] top-level scope at REPL[24]:1
```

I hope this is not a silly mistake on my side. I updated the code and restarted to copy-paste the first example...
