### 2021-04-15 #3053 @glwagner
@navidcy do you need to update the compat tag for `FourierFlows`? It's set to 0.6.16, but I think the reexport stuff was added to 0.6.17?

https://github.com/FourierFlows/GeophysicalFlows.jl/blob/e4f1ba3402eaa9982b62f4775bc58680290cf1a2/Project.toml#L24

https://github.com/FourierFlows/FourierFlows.jl/pull/271/files
