### 2021-04-15 e157b42
update FourierFlows compat
