### 2021-04-15 #3053 @navidcy
@pdebuyl, I pushed a patch v0.12.1. Could you try updating again and give it a go?
If you get an error then please paste here the output from
```julia
julia> using Pkg; Pkg.status()
```
and
```julia
julia> versioninfo()
```
