### 2021-04-16 a9ac7d9
drop in-house @has_cuda macro
