### 2021-04-16 #3053 @navidcy
Thanks. I’ll deal with these tomorrow.
@BrodiePearson is there a doi for your paper yet?
