### 2021-04-16 #3053 @navidcy
@pdebuyl, should the release that will be attached to the JOSS paper include the `paper` folder (e.g., merge https://github.com/FourierFlows/GeophysicalFlows.jl/tree/JOSS-paper to master and then make a release)?
