### 2021-04-16 #3053 @pdebuyl
@navidcy the paper can remain in a separate branch, see https://joss.readthedocs.io/en/latest/submitting.html#submission-requirements

Many authors let their paper in `paper/` or `doc/paper`, but it is not mandatory. For publishing purposes, JOSS stores the pdf and the metadata of the paper.
