### 2021-04-16 #3053 @BrodiePearson
@navidcy It has been assigned the following DOI: 10.1017/jfm.2021.247
Because it is in the proofing stage this DOI has not been activated yet, is it still allowed? I expect it to be published imminently, and I do not have a preprint.
