### 2021-04-16 #3053 @navidcy
I can add the DOI but I don't know if `whendon` will still complain since it's not active...

@BrodiePearson, any chance you push it onto arXiv so we cite that? Then, as soon as paper is published you can add the DOI and full publications details in the arXiv as metadata.
