+++ paper/paper.bib
@@ -16,7 +16,8 @@
   volume={5},
   number={53},
   pages={2018},
-  year={2020}
+  year={2020},
+	doi={10.21105/joss.02018}
 }
 
 @article{MAOOAM,
@@ -43,12 +44,12 @@
   author       = {Constantinou, N. C. and
                   Wagner, L. C. and
                   Palóczy, A.},
-  title        = {FourierFlows/FourierFlows.jl: v0.6.13},
+  title        = {FourierFlows/FourierFlows.jl: v0.6.17},
   year         = {2021},
   publisher    = {Zenodo},
-  version      = {v0.6.13},
-  doi          = {10.5281/zenodo.4460734},
-  url          = {https://doi.org/10.5281/zenodo.4460734}
+  version      = {v0.6.17},
+  doi          = {10.5281/zenodo.4686348},
+  url          = {https://doi.org/10.5281/zenodo.4686348}
 }
 
 @article{Thyng2016,
@@ -95,12 +96,13 @@
   url          = {https://doi.org/10.5281/zenodo.3551326}
 }
 
-@article{Pearson2021,
+@preprint{Pearson2021,
   title = {Advective structure functions in anisotropic two-dimensional turbulence},
   author = {Pearson, B. C. and Pearson, J. L. and Fox-Kemper, B.},
   journal = {J. Fluid Mech.},
   year = {2021},
-	note = {in press}
+	note = {in press},
+	doi = {10.1017/jfm.2021.247}
 }
 
 @article{Karrasch2020,
