### 2021-04-16 #3053 @navidcy
> Hi @navidcy
>
> I could test succesfully some of the examples, thanks.

great!

> A few extra comments before proceeding:
>
> * the zenodo citation for FourierFlows in the paper lists three authors, whereas the zenodo page lists five. Can you either edit the zenodo entry or the citation so that they match?

done

> * the Oceananigans.jl should have a doi

done

> * minor style suggestion "to grasp on new concepts and phenomena" -> "to grasp new concepts and phenomena"

done

> Is Version v0.12.1 10.5281/zenodo.4695260 the suitable code archive for the paper? We ask that the zenodo archive is title equally to the paper, can you edit the title to "GeophysicalFlows.jl: Solvers for geophysical fluid dynamics problems in periodic domains on CPUs & GPUs" ? The author list for the archive is not the same as the paper, I'll check in our documentation if this is an issue or not.

yes, that's a good version to associate with the JOSS paper. I've changed the Zenodo metadata to reflect the paper title and authors.

(The only outstanding remark is the citation to Pearson et al. 2021.)
