### 2021-04-19 #3053 @navidcy
> @navidcy thank you for the updates. Regarding the upcoming paper in J. Fluid Mech. how long until it public? If you don't want to wait I can check with the other editors how to handle such a reference.

@BrodiePearson any ETA on that?
