### 2021-04-19 5478bc2
update Pearson et al. bit entry
