+++ paper/paper.bib
@@ -97,11 +97,12 @@
 }
 
 @preprint{Pearson2021,
-  title = {Advective structure functions in anisotropic two-dimensional turbulence},
+  title = {Advective structure functions in anisotropic two-dimensional turbulence	},
   author = {Pearson, B. C. and Pearson, J. L. and Fox-Kemper, B.},
   journal = {J. Fluid Mech.},
   year = {2021},
-	note = {in press},
+	volume = {916},
+	pages = {A49},
 	doi = {10.1017/jfm.2021.247}
 }
 
