### 2021-04-19 22de1cd
update Pearson et al. bit entry
