+++ paper/paper.bib
@@ -99,7 +99,7 @@
 @preprint{Pearson2021,
   title = {Advective structure functions in anisotropic two-dimensional turbulence	},
   author = {Pearson, B. C. and Pearson, J. L. and Fox-Kemper, B.},
-  journal = {J. Fluid Mech.},
+  journal = {Journal of Fluid Mechanics},
   year = {2021},
 	volume = {916},
 	pages = {A49},
