### 2021-04-19 e6cd670
update Pearson et al. bit entry
