+++ paper/paper.bib
@@ -96,7 +96,7 @@
   url          = {https://doi.org/10.5281/zenodo.3551326}
 }
 
-@preprint{Pearson2021,
+@article{Pearson2021,
   title = {Advective structure functions in anisotropic two-dimensional turbulence	},
   author = {Pearson, B. C. and Pearson, J. L. and Fox-Kemper, B.},
   journal = {Journal of Fluid Mechanics},
