### 2021-04-19 #3053 @navidcy
@pdebuyl I just went through the proofs.

Looks good -- I just rephrased a the sentence to reflect that Pearson et al. 2021 is part of what "has been done" and not part of what's "Currently, `GeophysicalFlows.jl` is used for..." and pushed in `JOSS-paper` branch. (see https://github.com/FourierFlows/GeophysicalFlows.jl/pull/233/commits/38af08b1a4212563ccebb6909c8133f3bb01a9f0).

Can you update the pdf to reflect the latest version?
